LDE-714: moved tests for build_cube
